feat(cloud-sql/postgres): update to V2 sample
